:truck: Move model files. - Moved `./src/models/__init__.py` -> `./src/common/models/__init__.py`. - Moved `./src/models/auth.py` -> `./src/common/models/auth.py`. - Moved `./src/models/client.py` -> `./src/common/models/client.py`. - Moved `./src/models/message.py` -> `./src/common/models/message.py`. - Moved `./src/models/receipt.py` -> `./src/common/models/receipt.py`. - Moved `./src/models/responses.py` -> `./src/common/models/responses.py`. - Moved `./src/models/ticket.py` -> `./src/common/models/ticket.py`. - Moved `./src/models/user.py` -> `./src/common/models/user.py`. - Moved `./src/models/webhook.py` -> `./src/common/models/webhook.py`. - Deleted `./src/models/context.py`. - Deleted `./src/models/event.py`.
