+++ src/common/models/receipt.py
