--- src/models/__init__.py
@@ -1,29 +1,0 @@
-from src.models.auth import LoginSchema, RegisterSchema, TokenSchema
-from src.models.client import Client, ClientInsert
-from src.models.context import ContextType
-from src.models.event import Event
-from src.models.message import Message, MessageInsert
-from src.models.receipt import Receipt
-from src.models.responses import GenericResponse, create_response
-from src.models.ticket import Ticket
-from src.models.user import User, UserRole
-from src.models.webhook import TwilioWebHook
-
-__all__ = [
-    "Client",
-    "ClientInsert",
-    "ContextType",
-    "create_response",
-    "Event",
-    "GenericResponse",
-    "LoginSchema",
-    "Message",
-    "MessageInsert",
-    "Receipt",
-    "RegisterSchema",
-    "Ticket",
-    "TokenSchema",
-    "TwilioWebHook",
-    "User",
-    "UserRole",
-]
