--- src/models/context.py
@@ -1,19 +1,0 @@
-from typing import TypedDict
-
-from src.models.event import Event
-from src.services.chat_api import ChatApiService
-from src.services.saragurosnet import SaragurosNetService
-from src.services.twilio import TwilioService
-
-
-# TODO: Review with Aminael if this value types are correct
-class ContextType(TypedDict):
-    """ContextType class to handle the context of the application."""
-
-    EVENT_TWILIO: Event
-    DATA_USER_ID: int
-    DATA_USER_CI: str | None
-    DATA_LAST_STATE: str
-    SERVICE_API: ChatApiService
-    SERVICE_SARAGUROS: SaragurosNetService
-    SERVICE_TWILIO: TwilioService
